des trucs

+++ core/defines.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <iostream>
-
+#include <cmath>
 
 // Definitions
 
@@ -67,5 +67,5 @@ using u16 = unsigned short;
 using u8 = unsigned char;
 
 using ParticleId = u32;
-using Pixel = u8;
+using Pixel = float;
